requirements.txt
requests
python-dotenv
beautifulsoup4
ipumspy
pandas
numpy
